refactor: expression.tsをlibディレクトリに移動
