+++ app/components/Display/index.tsx
@@ -1,4 +1,4 @@
-import { Expression } from "@/app/expression";
+import { Expression } from "@/lib/expression";
 import styles from "./Display.module.css";
 
 export default function Display({ expression }: { expression: Expression }) {
