+++ lib/expression.ts
